Refactored index.js: moved transformFile to own file
