+++ test/testdata/input/multiline_file.js
@@ -3,9 +3,7 @@ const value = 'value';
 const badValue = 1;
 shouldBeEqualToString("value", 'value');
 shouldNotBe("value.length", "badValue");
-let testObject = {
-  property: true
-};
+let testObject = { property: true };
 
 debug('tested value');
 description('This part tests testObject but this is a bad description!');
