+++ test/testdata/reference/transformed_file-list-test.html
@@ -6,7 +6,7 @@
 <!DOCTYPE html>
 <html>
 <head>
-<title>Test the attribute of FileList</title>
+<title>Test the attribute of FileList.</title>
 <script src="../../resources/testharness.js"></script>
 <script src="../../resources/testharnessreport.js"></script>
 </head>
@@ -15,14 +15,13 @@
 <div id="description"></div>
 <div id="console"></div>
 <script>
-setup({
-  single_test: true,
-  explicit_done: false
-});
+setup({ single_test: true, explicit_done: false });
+
 
 console.log("Start");
 
-function onInputFileChange(files) {
+function onInputFileChange(files)
+{
   window.files = files;
   assert_true(files instanceof FileList);
   assert_equals(files.length, 2);
@@ -45,7 +44,6 @@ function onInputFileChange(files) {
 eventSender.beginDragWithFiles(["resources/UTF8.txt", "resources/UTF8-2.txt"]);
 eventSender.mouseMoveTo(10, 10);
 eventSender.mouseUp();
-
 </script>
 </body>
 </html>
