+++ test/testdata/reference/transformed_multiline_file.js
@@ -1,14 +1,11 @@
-setup({
-  single_test: true,
-  explicit_done: false
-});
+setup({ single_test: true, explicit_done: false });
 const value = 'value';
 const badValue = 1;
 assert_equals(value, 'value');
 assert_not_equals(value.length, badValue);
-let testObject = {
-  property: true
-};
+let testObject = { property: true };
+
 console.log('tested value');
+
 assert_not_equals(testObject.property, undefined);
 assert_true(testObject.property);
