+++ test/testdata/reference/transformed_short_file.js
@@ -1,5 +1,2 @@
-setup({
-  single_test: true,
-  explicit_done: false
-});
+setup({ single_test: true, explicit_done: false });
 assert_equals(typeof nameVar, 'string');
